Integrate 'new comment' and 'list comments' servlet into one data servlet
