--- portfolio/src/main/java/com/google/sps/servlets/NewCommentServlet.java
@@ -1,67 +1,0 @@
-// Copyright 2019 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package com.google.sps.servlets;
-
-import com.google.appengine.api.datastore.DatastoreService;
-import com.google.appengine.api.datastore.DatastoreServiceFactory;
-import com.google.appengine.api.datastore.Entity;
-import java.io.IOException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import com.google.gson.Gson;
-import java.util.ArrayList;
-
-/** Servlet that stores new comments. */
-@WebServlet("/new-data")
-public class NewCommentServlet extends HttpServlet {
-
-  private static final DatastoreService datastore = DatastoreServiceFactory.getDatastoreService();
-    
-  @Override
-  public void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
-    // Retrieve information from the form and add timestamp
-    final String name = getParameterWithDefault(request, "user-name", "Anonymous");
-    final String job = getParameterWithDefault(request, "jobs", "Other");
-    final String comment = getParameterWithDefault(request, "visitor-comment", "");
-    final long timestamp = System.currentTimeMillis();
-
-    // Store the information if comment is non-empty
-    if (!comment.isEmpty()) {
-      Entity messageEntity = new Entity("Message");
-      messageEntity.setProperty("name", name);
-      messageEntity.setProperty("job", job);
-      messageEntity.setProperty("comment", comment);
-      messageEntity.setProperty("timestamp", timestamp);
-      datastore.put(messageEntity);
-    }
-
-    // Redirect back to the HTML page.
-    response.sendRedirect("/index.html");
-  }
-
-  /**
-   * @return the request parameter, or the default value if the parameter
-   *         was not specified by the client
-   */
-  private String getParameterWithDefault(HttpServletRequest request, String name, String defaultValue) {
-    String value = request.getParameter(name);
-    if (value == null) {
-      return defaultValue;
-    }
-    return value;
-  }
-}
